Bugfixes and better handling of Paragraphs and List Styles #time 2h
